requirements.txt
# Core LangChain
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langchain-core>=0.3.0
langgraph>=0.2.0
# MCP
fastmcp>=2.0.0
# ACP
acp-sdk>=0.4.0
# Document processing
pypdf>=4.0.0
# Web search and scraping
duckduckgo-search>=6.0.0
trafilatura>=2.0.0
httpx>=0.27.0
# Vector store
faiss-cpu>=1.8.0
# BM25 search
rank-bm25>=0.2.2
nltk>=3.8.0
# Reranking
sentence-transformers>=3.0.0
# Pydantic
pydantic>=2.0.0
# Utilities
python-dotenv>=1.0.0
numpy>=1.26.0
langchain-text-splitters